% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_default}
\alias{config_default}
\title{Define the parameters specific to each differential analysis method.}
\usage{
config_default(
  alpha = 0.1,
  minreads = 10,
  ncores = 4,
  run_deseq2 = T,
  run_edgeR = T,
  run_limma = T,
  ...
)
}
\arguments{
\item{alpha}{fdr level when adjusting for multiple testing.}

\item{minreads}{the minimum number of reads a variable must have to be differentially analyzed.}

\item{ncores}{number of cores available for doing parallel computations. Used in DESeq.}

\item{run_deseq2}{boolean to choose to run DESeq2.}

\item{run_edgeR}{boolean to choose to run edgeR.}

\item{run_limma}{boolean to choose to run limma.}

\item{...}{extra parameters added to the configuration list}
}
\description{
Define the parameters specific to each differential analysis method.
}
\author{
Yoann Pradat
}
