% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{get_contrast_condition_B_minus_A_factors_interaction}
\alias{get_contrast_condition_B_minus_A_factors_interaction}
\title{Build a list of character contrasts of B-A for an interacting variable.}
\usage{
get_contrast_condition_B_minus_A_factors_interaction(
  levels_cond,
  levels_inte,
  cond_name = "condition",
  inte_name = "interaction",
  level_B_cond = NULL,
  level_A_cond = NULL,
  cond_main_effect = T,
  inte_main_effect = F
)
}
\arguments{
\item{levels_cond}{the factor levels of the condition variable}

\item{levels_inte}{the factor levels of the interaction variable}

\item{cond_name}{the name of the condition variable}

\item{inte_name}{the name of the interaction variable}

\item{level_B_cond}{the B level to be contrasted with A}

\item{level_A_cond}{the A level to be contrasted with B}

\item{cond_main_effect}{is the condition a main effect}

\item{inte_main_effect}{is the interaction a main effect}
}
\description{
Build a list of character contrasts of B-A for an interacting variable.
}
