% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{get_contrast_vector}
\alias{get_contrast_vector}
\title{Produce a contrast vector}
\usage{
get_contrast_vector(contrast, design, data)
}
\arguments{
\item{contrast}{a character vector specifying the contrast beween refined beta names (see
\code{\link{refine_beta_names}}}

\item{design}{a formula specifying the design}

\item{data}{a data.frame used for building the model matrix}
}
\value{
a character vector of refined beta names
}
\description{
Produce a contrast vector
}
\references{
internal
}
\author{
Yoann Pradat
}
