% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_to_limma}
\alias{load_to_limma}
\title{Transform \code{SummarizedExperiment} object to ????}
\usage{
load_to_limma(object)
}
\arguments{
\item{object}{a \code{SummarizedExperiment} object}
}
\description{
Transform \code{SummarizedExperiment} object to ????
}
\author{
Yoann Pradat
}
