% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opts_diffexp}
\alias{opts_diffexp}
\title{Define the parameters specific to each differential analysis method.}
\usage{
opts_diffexp(
  alpha = 0.1,
  ncores = 6,
  save_table = T,
  only_significant = T,
  folder_results = "./results",
  use_deseq2 = T,
  use_edgeR = F,
  use_limma = F,
  ...
)
}
\arguments{
\item{alpha}{fdr level when adjusting for multiple testing.}

\item{ncores}{number of cores available for doing parallel computations. Used in DESeq.}

\item{save_table}{boolean to decide whether to save tables in txt files or not}

\item{only_significant}{boolean to decide whether only significant (FDR) variables are kept in the results tables
or not}

\item{use_deseq2}{boolean to choose to run DESeq2.}

\item{use_edgeR}{boolean to choose to run edgeR.}

\item{use_limma}{boolean to choose to run limma.}

\item{...}{extra parameters added to the configuration list}
}
\description{
Define the parameters specific to each differential analysis method.
}
\author{
Yoann Pradat
}
