% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opts_prepro}
\alias{opts_prepro}
\title{Define the options for \code{\link{preprocess_object}}}
\usage{
opts_prepro(
  design = NULL,
  min_count = 0,
  min_total_count = 15,
  large_n = 10,
  min_prop = 0.7,
  norm_factors_method = c("TMM", "TMMwsp", "RLE", "upperquartile", "none"),
  ...
)
}
\arguments{
\item{design}{a formula object specifying the design matrix}

\item{min_count}{\code{min.total.count} param of \code{\link[edgeR]{filterByExpr}}}

\item{min_total_count}{\code{min.count} param of \code{\link[edgeR]{filterByExpr}}}

\item{large_n}{\code{large.n} param of \code{\link[edgeR]{filterByExpr}}}

\item{min_prop}{\code{min.prop} param of \code{\link[edgeR]{filterByExpr}}}

\item{norm_factors_method}{\code{method} argument of \code{\link[edgeR]{calcNormFactors}}.}

\item{...}{extra parameters added to the options list}
}
\value{
a \code{SummarizedExperiment} object
}
\description{
Define the options for \code{\link{preprocess_object}}
}
\author{
Yoann Pradat
}
