% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dispersion_deseq2}
\alias{plot_dispersion_deseq2}
\title{Plot dispersion estimates from DESeq2}
\usage{
plot_dispersion_deseq2(dds, filepath, title, subtitle = NULL)
}
\arguments{
\item{dds}{a \code{DESeqDataSet} object}

\item{filepath}{a character specifying the path to the plot file}

\item{title}{a character}

\item{subtitle}{(optional) a character}
}
\description{
Plot dispersion estimates from DESeq2
}
\keyword{internal}
