% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dispersion_ql_edgeR}
\alias{plot_dispersion_ql_edgeR}
\title{Plot dispersion estimates from edgeR}
\usage{
plot_dispersion_ql_edgeR(glm_fit, filepath, title, subtitle = NULL)
}
\arguments{
\item{glm_fit}{a \code{DGEGLM} object}

\item{filepath}{a character specifying the path to the plot file}

\item{title}{a character}

\item{subtitle}{(optional) a character}
}
\description{
Plot dispersion estimates from edgeR
}
\keyword{internal}
