% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_MA_deseq2}
\alias{plot_MA_deseq2}
\title{Plot MA from DESeq2}
\usage{
plot_MA_deseq2(
  resLFC_MLE,
  resLFC_shrink = NULL,
  filepath,
  title,
  subtitle = NULL,
  alpha = 0.1,
  ylim = c(-2, 2)
)
}
\arguments{
\item{resLFC_MLE}{a \code{DESeqResults} object}

\item{resLFC_shrink}{(optional) a \code{DESeqResults} object}

\item{filepath}{a character specifying the path to the plot file}

\item{title}{a character}

\item{subtitle}{(optional) a character}

\item{alpha}{a float passed to \code{\link[DESeq2]{plotMA}}}

\item{ylim}{a length-2 float vector passed to \code{\link[DESeq2]{plotMA}}}
}
\description{
Plot MA from DESeq2
}
\keyword{internal}
