% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{refine_beta_names}
\alias{refine_beta_names}
\title{Add references levels to beta names created by \code{model.matrix}}
\usage{
refine_beta_names(design, data)
}
\arguments{
\item{design}{a formula specifying the design}

\item{data}{a data.frame used for building the model matrix}
}
\value{
a character vector of refined beta names
}
\description{
Add references levels to beta names created by \code{model.matrix}
}
\references{
internal
}
\author{
Yoann Pradat
}
