% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_edgeR}
\alias{run_edgeR}
\title{Run edgeR algorithm.}
\usage{
run_edgeR(object, design = NULL, opts_algo, opts_comm)
}
\description{
Run edgeR algorithm.
}
\references{
Robinson MD, McCarthy DJ, Smyth GK (2010). “edgeR: a Bioconductor package for differential expression analysis of
digital gene expression data.” Bioinformatics, 26(1), 139-140. \url{https://doi.org/10.1093/bioinformatics/btp616}

McCarthy DJ, Chen Y, Smyth GK (2012). “Differential expression analysis of multifactor RNA-Seq experiments with
respect to biological variation.” Nucleic Acids Research, 40(10), 4288-4297. \url{https://doi.org/10.1093/nar/gks042}
}
\author{
Yoann Pradat
}
