% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_limma}
\alias{run_limma}
\title{Run limma algorithm.}
\usage{
run_limma(object)
}
\description{
Run limma algorithm.
}
\references{
Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK (2015). “limma powers differential expression analyses for
RNA-sequencing and microarray studies.” Nucleic Acids Research, 43(7), e47. \url{https://doi.org/10.1093/nar/gkv007}

Law, C.W., Chen, Y., Shi, W. et al. voom: precision weights unlock linear model analysis tools for RNA-seq read
counts. Genome Biol 15, R29 (2014). \url{https://doi.org/10.1186/gb-2014-15-2-r29}
}
\author{
Yoann Pradat
}
