% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_brca-data.R
\docType{data}
\name{tcga_brca}
\alias{tcga_brca}
\title{RNA-seq and clinical data for TCGA BRCA Stage I patients.}
\format{
A \code{SummarizedExperiment} object with 58,288 genes and 98 samples.
\describe{
\item{assays}{A \code{SimpleList} with just one "counts" matrix}
\item{colData}{A \code{DFrame} with 98 rows and 8 columns}
\item{rowData}{A \code{DFrame} with 58,288 rows and 2 columns}
\item{metadata}{An empty list}
\item{elementMetadata}{Alias for \code{rowData}}
}
}
\source{
RNA-seq source: \url{https://stanfordmedicine.app.box.com/s/lu703xuaulfz02vgd2lunxnvt4mfvo3q}. cBioPortal
BRCA Pan-cancer Atlas \url{https://www.cbioportal.org/study/summary?id=brca_tcga_pan_can_atlas_2018}. GDC legacy archive
clinical BRCA \url{https://portal.gdc.cancer.gov/legacy-archive/files/735bc5ff-86d1-421a-8693-6e6f92055563}
}
\usage{
data(tcga_brca)
}
\description{
Data from the reanalysis of RNA-seq files by Zheng et al. 2019. The clinical data were taken from the GDC data
portal and the cBio data portal. Both the RNA-seq data and the clinical data are public access. The samples were
chosen so that
\enumerate{
\item all samples are Stage I
\item all samples are Female
\item all samples have complete clinical data
}
}
\details{
This selection was performed only to keep this example data set small and may not be suitable for meaningful
analyses.
}
\examples{
data(tcga_brca)
}
\references{
Hong Zheng, Kevin Brennan, Mikel Hernaez, Olivier Gevaert, Benchmark of long non-coding RNA quantification for RNA
sequencing of cancer samples, GigaScience, Volume 8, Issue 12, December 2019, giz145,
\url{https://doi.org/10.1093/gigascience/giz145}
}
\keyword{datasets}
