% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{update_table}
\alias{update_table}
\title{Update an existing table with a new table.}
\usage{
update_table(tab, tab_new, tags)
}
\arguments{
\item{tab}{a data.frame}

\item{tab_new}{a data.frame}

\item{tags}{named list of tags values to decide between merging and updating}
}
\value{
a data.frame
}
\description{
Update an existing table with a new table.
}
\author{
Yoann Pradat
}
\keyword{internal}
