% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_to_deseq2}
\alias{load_to_deseq2}
\title{Transform \code{SummarizedExperiment} object to \code{\link[DESeq2]{DESeq2DataSet}} object.}
\usage{
load_to_deseq2(object, design = NULL)
}
\arguments{
\item{object}{a \code{SummarizedExperiment} object}

\item{design}{a formula or matrix that expresses how the counts for each variable depend on the variables in
\code{colData}. See \code{\link[DESeq2]{DESeqDataSet}}.}
}
\value{
a \code{DESeq2DataSet} object
}
\description{
Transform \code{SummarizedExperiment} object to \code{\link[DESeq2]{DESeq2DataSet}} object.
}
\author{
Yoann Pradat
}
