% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_to_edgeR}
\alias{load_to_edgeR}
\title{Transform \code{SummarizedExperiment} object to \code{\link[edgeR]{DGEList}} object.}
\usage{
load_to_edgeR(object)
}
\arguments{
\item{object}{a \code{SummarizedExperiment} object}
}
\description{
Transform \code{SummarizedExperiment} object to \code{\link[edgeR]{DGEList}} object.
}
\author{
Yoann Pradat
}
