% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_object}
\alias{preprocess_object}
\title{Perform preprocessing steps.}
\usage{
preprocess_object(object, opts)
}
\arguments{
\item{object}{a \code{SummarizedExperiment} object}

\item{opts}{a named list of options. See \code{\link{opts_prepro}}}
}
\value{
a \code{SummarizedExperiment} object
}
\description{
Perform preprocessing steps.
}
\author{
Yoann Pradat
}
