% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run DESEQ2 algorithm.}
\usage{
run_deseq2(
  object,
  design = NULL,
  lfcShrink_type = "apeglm",
  contrasts,
  ncores = 1,
  alpha = 0.1,
  save_table = F,
  only_significant = T
)
}
\arguments{
\item{object}{a \code{SummarizedExperiment} object}

\item{design}{a formula specifying the design for the modelMatrix of DESeq2. Any variable appearing should be present
in the colData of object.}

\item{lfcShrink_type}{(optional) a character vector specifying type parameter of \code{lfcShrink}}

\item{contrasts}{(optional) a character vector specifying the contrasts (one or multiple beta coefficient) to be used
for making tests and building results table. If set to NULL, one table for each element of resultsNames(dds_main)
will be built.}

\item{ncores}{(optional) an integer specifying the number of cores available for running DESeq2}

\item{alpha}{(optional) the FDR rate level for multiple testing}

\item{save_table}{(optional) whether the table of resutls should be written in a text file or not.}

\item{only_significant}{(optional) whether the table of resutls should include only the variables with padj < alpha
not.}
}
\value{
dataframe with results
}
\description{
Run DESEQ2 algorithm.
}
\references{
Love, M.I., Huber, W., Anders, S. (2014) Moderated estimation of fold change and dispersion for RNA-seq data with
DESeq2. Genome Biology, 15:550. \url{https://doi.org/10.1186/s13059-014-0550-8}
}
\author{
Yoann Pradat
}
