% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run DESEQ2 algorithm.}
\usage{
run_deseq2(object, design = NULL, contrasts, opts_algo, opts_comm)
}
\arguments{
\item{object}{a \code{SummarizedExperiment} object}

\item{design}{a formula specifying the design for the model matrix of DESeq2. Any variable appearing should be present
in the \code{colData} of object}

\item{contrasts}{(optional) a character vector specifying the contrasts (one or multiple beta coefficient) to be used
for making tests and building results table. If set to NULL, one table for each element of \code{resultsNames(dds)}
will be built}

\item{opts_algo}{a named list of options passed to \code{DESeq}, \code{results} and \code{lfcShrink} functions. See
\code{\link[DESeq2]{DEseq}}, \code{\link[DESeq2]{results}}, \code{\link[DESeq2]{lfcShrink}}}
}
\value{
dataframe with results
}
\description{
Run DESEQ2 algorithm.
}
\references{
Love, M.I., Huber, W., Anders, S. (2014) Moderated estimation of fold change and dispersion for RNA-seq data with
DESeq2. Genome Biology, 15:550. \url{https://doi.org/10.1186/s13059-014-0550-8}
}
\author{
Yoann Pradat
}
