% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_edgeR}
\alias{run_edgeR}
\title{Run edgeR algorithm.}
\usage{
run_edgeR(object, design = NULL, contrasts, opts_algo, opts_comm)
}
\arguments{
\item{object}{a \code{SummarizedExperiment} object}

\item{design}{a formula specifying the design for the model matrix of DESeq2. Any variable appearing should be
present in the \code{colData} of object}

\item{contrasts}{a character vector specifying the contrasts (one or multiple beta coefficient) to
be used for making tests and building results table.}

\item{opts_algo}{a named list of options specific to edgeR}

\item{opts_comm}{a named list of options common to all methods}
}
\value{
a dataframe of results
}
\description{
Run edgeR algorithm.
}
\references{
Robinson MD, McCarthy DJ, Smyth GK (2010). “edgeR: a Bioconductor package for differential expression analysis of
digital gene expression data.” Bioinformatics, 26(1), 139-140. \url{https://doi.org/10.1093/bioinformatics/btp616}

McCarthy DJ, Chen Y, Smyth GK (2012). “Differential expression analysis of multifactor RNA-Seq experiments with
respect to biological variation.” Nucleic Acids Research, 40(10), 4288-4297. \url{https://doi.org/10.1093/nar/gks042}
}
\author{
Yoann Pradat
}
